Add V1.1: Epipolar Cross-Attention + Fourier PE architecture
Replace Inverse Splat with geometry-aware epipolar cross-attention and
upgrade Plucker MLP to Fourier-encoded PE system. Remove image
reconstruction stage; depth becomes lightweight 1x1 Conv auxiliary loss.
Single-stage training (no curriculum) with both label models loaded
simultaneously.

New files:
- train/models/geometric_pe.py (FourierRayPE, Target3DPE, CameraIDEmbedding)
- train/models/epipolar_attention.py (EpipolarCrossAttention)
- train/models/image_adaptor_v1_1.py (V1.1 model wrapper)
- train/train_v1_1.py (simplified training script)
- config/v1_1.yaml (V1.1 config)
- sketch/V1.1/ (design documents)

V1 fixes:
- Fix Qwen3-VL visual encoder access pattern in OnlineLabelProvider
- Add merger step for proper token extraction (720→180 tokens)
- Cast tokens to float32 for stable loss computation
- Add feature map PCA/mean/std visualization to TensorBoard
- Fix label generation order to avoid VRAM ping-pong

+++ train/losses/token_loss.py
@@ -132,10 +132,10 @@ class TokenLoss(nn.Module):
                 p = pred[cam]
                 t = target[cam]
 
-                # Truncate/pad to match sizes
+                # Truncate/pad to match sizes, cast to float32 for stable loss
                 min_tokens = min(p.shape[1], t.shape[1])
-                p = p[:, :min_tokens]
-                t = t[:, :min_tokens]
+                p = p[:, :min_tokens].float()
+                t = t[:, :min_tokens].float()
 
                 total_mse = total_mse + self.mse_loss(p, t)
                 cos = self.cos_sim(p, t).mean()
